Fix location of data volume on nycpad_normalizer
The `docker-compose.yml` mounts the data volume on `/data`, but the actual container writes its data (which is later retrieved by the `nycpad` container for importing) to `/usr/local/src/scripts/data`.  This changes the location of the mount to match where data is written, so that everything works as expected when `pelias import nycpad` is run.